Fix #69: Remove VideoPlaybackQuality from MSE (#135)
Removes the VideoPlaybackQuality object and HTMLVideoElement extension
method getVideoPlaybackQuality(), but retains brief mention of them in a
link to the July 5, 2016 MSE Candidate Recommendation document draft in
a non-normative appendix. Includes also an informative
reference link to the current WICG Media Playback Quality effort in that
appendix.

Also updates the "Introduction" section to be explicitly non-normative.