Remove appendStream() and related spec text from MSE v1
I believe we have editors' consensus to remove the appendStream feature
(there's a CfC currently in progress) from MSE v1.

This commit removes that feature. We can revert it if the CfC for the
removal fails, or if we wish to refine the spec and implementation in
VNext, in the appropriate branch.

See also
https://github.com/w3c/media-source/issues/14#issuecomment-238363951