Use absolute URL for MSE in xref script, and fix broken fragments
The `media-source.js` script is used to manage cross-references in MSE specs.
To be able to reference this script from dedicated repositories for byte stream
formats and registry specifications (see #239), the script must use an absolute
URL for the ED of the main Media Source Extensions spec.

The update also fixes a few broken fragments (naming rules have changed in
Respec in the meantime).