Switch to Process 2020 (#263)
This update makes a (somewhat) minimal set of changes to have the Status of
this Document section say that the document is governed by Process 2020.

In theory, that should be as simple as switching to the right profile of ReSpec.
In practice, that version of ReSpec changed a few things on the way definitions
are handled, requiring a few other updates to definitions and links not to get
linking errors and warnings from ReSpec.

One of these updates is using the right shorthand to link back occurrences of
readyState (the attribute) and ReadyState (the enum). ReSpec now supports case
sensitivity for IDL terms, and the workaround implemented in media-source.js was
not perfect in any case (in the published Rec, an instance of "ReadyState" links
back to the attribute definition in the IDL block instead of to the definition
of the enum).