Modernize var syntax, including type info where known
Updates `<var>...</var>` to instead use modern respec `|...|`, and in
most cases, on first definition of a variable, includes type of variable
in a suffix to `...` in the definition.

Links IDL method arguments in the definition to the argument variable's
detail by also wrapping those argument names in `|...|`.

Includes a couple other minor changes: uses `minus` instead of `-` in
algorithms in a couple places.

Replaces a couple instances of "source buffer" with "{{SourceBuffer}}"
for improved linkage.

Note, the MSE-in-Workers initial draft PR (#282) contained some similar
changes to lines only relevant to that draft, requiring manual rebase of
this onto it once that landed.