Use internal slot for SB `generate timestamps flag`
Also adds text in the definition that mentions that `changeType()`
updates `generate timestamps flag`.

Note, this dfn is exported, so any previous external references to it via
[=generate timestamps flag=] might need to be changed to
{{SourceBuffer/[[generate timestamps flag]]}} wherever it is referenced.

In particular, media-source.js is updated to (hopefully) enable correct
fragment reference for this definition from MSE BSF Registry ED. The
longer-term fix will be to update that registry's respec directly to
use xref and the linkage, above.