Update SOTD to highlight substantive changes since REC
See discussion in #294.

This updates the Status of this Document section in preparation to publication
as First Public Working Draft, highlighting substantive changes made to the spec
and pointing people at the list of commits for details.

(Also fixes a typo in the way the File API was referenced to fix the incorrect
duplicate FILEAPI entry that appeared in the informative references at the end
of the spec)