Fix broken refs to HTML spec, adjust settings for FPWD
The spec referenced a number of terms in the HTML spec using links to the
superseded HTML 5.1 spec. That spec actually redirects to the WHATWG HTML LS
spec, which is good. Problem is a good chunk of anchors have been modified in
the HTML spec and are no longer valid.

This update fixes all broken fragments using a combination of:
- "xref" terms, automatically handled by ReSpec, when possible (i.e. when the
HTML spec exports them).
- fragment updates in media-source.js when the spec needs to reference
non-exported terms. Longer-term, we'll need to work with HTML editors to make
sure that the spec properly exports the terms that we need to reference.

These updates are notably needed to pass pubrules for publication as First
Public Working Draft.

One note on xref terms: the HTML spec no longer has the concept of "initial
playback position". The concept seems to have been merged with "current playback
position".

This update also includes a couple of other pubrule changes:
- Required editor ID was missing for Mark.
- Link to Can I Use was semi-broken (they still managed to redirect the request
but a more direct https://caniuse.com/mediasource can be used)