Fix broken refs to HTML spec, adjust settings for FPWD (#300)
* Fix broken refs to HTML spec, adjust settings for FPWD

The spec referenced a number of terms in the HTML spec using links to the
superseded HTML 5.1 spec. That spec actually redirects to the WHATWG HTML LS
spec, which is good. Problem is a good chunk of anchors have been modified in
the HTML spec and are no longer valid.

This update fixes all broken fragments using a combination of:
- "xref" terms, automatically handled by ReSpec, when possible (i.e. when the
HTML spec exports them).
- fragment updates in media-source.js when the spec needs to reference
non-exported terms. Longer-term, we'll need to work with HTML editors to make
sure that the spec properly exports the terms that we need to reference.

These updates are notably needed to pass pubrules for publication as First
Public Working Draft.

One note on xref terms: the HTML spec no longer has the concept of "initial
playback position". The concept seems to have been merged with "current playback
position".

This update also includes a couple of other pubrule changes:
- Required editor ID was missing for Mark.
- Link to Can I Use was semi-broken (they still managed to redirect the request
but a more direct https://caniuse.com/mediasource can be used)

* Drop link to implementation report

The implementation report document was for the REC, is likely outdated and is
no longer maintained. Group will need to create a new one over time (or update
that one) but it seems better not to link to it for the time being.

* Update shortname to media-source-2

Spec cannot keep the `media-source` shortname, because a REC already exists with
that shortname, see:
https://lists.w3.org/Archives/Public/public-media-wg/2021Sep/0011.html

Idea is to use a leveled shortname instead.

Note that this update does not use the `level` ReSpec setting on purpose to
keep the title intact ("Level 2" would be added to the title otherwise).

* Add missing "initial" adjective