MSE-for-WebCodecs: More partial updates
Adds TODO items to track more work necessary in this draft.
Fixes the addSourceBuffer parameter type to be TypeOrConfig.
References a not-yet-defined internal slot that tracks the pending
append chunks promise, and adds steps to reject that promise on abort()
or removeSourceBuffer(), if necessary.
Updates various places that describe BSF to also mention the possibility
that a SourceBuffer may instead be configured to expect WebCodecs
encoded chunks appends.
Adds a note to isTypeSupported() for how to use the WebCodecs
isConfigSupported() methods to proactively check support for buffering
encoded chunks into a SourceBuffer.