[CI job] Swap steps to fix publication to gh-pages (#313)
Publication to `gh-pages` currently fails because the w3c/spec-prod action
cannot find the `.git/config` file when it looks for it. The action runs
smoothly on a number of other repositories with very similar settings, and
manages to publish the spec to /TR without any problem. One explanation could
be that the step that copies static file somehow modifies the local state
(and drops the `.git` folder).

This update swaps the steps to run the spec generation job first and deployment
of static files last.