Change MediaSource handle getter value to null from exception
Due to compatibility issues with existing web app libraries that
enumerate MediaSource attributes, throwing an exception on attempted
read of a MediaSource object's handle attribute was unfortunately a
breaking change.

This change updates the spec to instead make the MediaSource handle
attribute readonly [SameObject] getter return null instead of throwing
an exception if the implementation doesn't support creating a
MediaSourceHandle for that MediaSource object instance.