Restrict handle getter visibility to dedicated worker
Removes the MediaSource handle attribute visibility from the main Window
context, leaving it visible only in dedicated worker contexts. Also
removes the conditional throwing of NotSupportedError exceptions on
getting this attribute's value if the implementation doesn't support
that.

Overall, this change is to undo the backwards incompatibility issue
found when older versions of MSE web app libraries enumerated main
thread MediaSource attribute values, hitting exception in the Chromium
feature implementation through milestone 107. This change replaces that
with simpler and safer visibility restriction of this getter to just the
context where it must be supported if the implementation supports
MSE-in-Workers.

Note, this change replaces the alternative fix discussed in
w3c/media-source PR #316.