Fix broken SOTD structure and duplicate text, and 2 typos
The Status of this Document section contained the list of changes twice. Also,
this section is not a good one to add subsections as ReSpec adds boilerplate
text after the custom text, which then appears to be in the subsection. This
update drops the subsection. It may still be a good idea to replace that text
with a changelog section, this is just intended as a quick fix!

`ManagedSourceBuffer` was not properly enclosed in `{{}}` in that text and did
not render well as a result.

Similarly, the note that discusses eviction reasons mentioned an event named
`bufferedchangeevent`, instead of `bufferedchange`.